Move domain and validation to new api module
