+++ p2panda-rs/src/api/mod.rs
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+mod domain;
+mod errors;
+pub mod validation;
+
+pub use domain::{next_args, publish};
+pub use errors::{DomainError, ValidationError};
