+++ p2panda-rs/src/lib.rs
@@ -65,6 +65,8 @@
 #[cfg(any(feature = "test-utils", test))]
 use rstest_reuse;
 
+#[cfg(any(feature = "storage-provider", test))]
+pub mod api;
 pub mod document;
 pub mod entry;
 pub mod graph;
