+++ p2panda-rs/src/test_utils/memory_store/mod.rs
@@ -5,13 +5,10 @@
 //! - `MemoryStore` implementation of all storage provider traits
 //! - `domain` and `validation` modules for publishing and validating entries and operations
 //! - helpers for populating a store with test data
-pub mod domain;
-pub mod errors;
 pub mod helpers;
 mod provider;
 mod stores;
 mod types;
-pub mod validation;
 
 pub use provider::MemoryStore;
 pub use types::{PublishedOperation, StorageEntry};
